% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrBoot.R
\name{mlrBoot}
\alias{mlrBoot}
\title{Bootstrap analysis of a multiple linear regression model}
\usage{
mlrBoot(data, iter, alpha = 0.05)
}
\arguments{
\item{data}{, a data frame with first column the dependent variable and all other columns predictors}

\item{iter}{, number of bootstrap iterations}

\item{alpha}{, confidence level}
}
\value{
, command line statistics and plots
}
\description{
Bootstrap analysis of a multiple linear regression model
}
\examples{
require(datasets)

iris<-as.data.frame(datasets::iris)
iris<-within(iris, {
Setosa<-ifelse(Species=="setosa", 1, 0)
Versicolor<-ifelse(Species=="versicolor", 1, 0)
rm(Species)
})

mlrBoot(iris, 1000)

}
