% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMLR.R
\name{shinyMLR}
\alias{shinyMLR}
\title{Interactive Multiple Linear Model

This function launches a shiny app to interact with a multiple linear model. There are _
datasets that can be explored. The app allows the user to view different metrics of
model validity through statistics and plots. The app also produces three different
types of estimates and confidence intervals: normal multiple linear regression estimates,
bootstrapped estimates, and bayesian estimates. The different estimates and intervals
can be compared as tables or by a visual plot.}
\usage{
shinyMLR()
}
\value{

}
\description{
Interactive Multiple Linear Model

This function launches a shiny app to interact with a multiple linear model. There are _
datasets that can be explored. The app allows the user to view different metrics of
model validity through statistics and plots. The app also produces three different
types of estimates and confidence intervals: normal multiple linear regression estimates,
bootstrapped estimates, and bayesian estimates. The different estimates and intervals
can be compared as tables or by a visual plot.
}
\examples{
\dontrun{ shinyMLR()}

}
