% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valCheck.R
\name{valCheck}
\alias{valCheck}
\title{Check the validity of a linear model}
\usage{
valCheck(lm, series = FALSE)
}
\arguments{
\item{lm}{a linear model}

\item{series}{a boolean variable to indicate whether linear model has a time component}
}
\value{
command line statistics and plots
}
\description{
Check the validity of a linear model
}
\examples{

iris<-as.data.frame(datasets::iris)
iris<-within(iris, {
Setosa<-ifelse(Species=="setosa", 1, 0)
Versicolor<-ifelse(Species=="versicolor", 1, 0)
rm(Species)
})

irislm<-lm(Sepal.Length~., data=iris)
valCheck(irislm)

}
