% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valCheck.R
\name{valCheck}
\alias{valCheck}
\title{Check Validity of Linear Model

This function will help determine the validity of a linear model.
The function produces statistics and or plots to assess
equality of variance, normality, multicollinearity, and outliers.
If \code{series=TRUE} then the function will also print a statistic
for independence. See vignette MLReg20_Explanation_And_Theory Section 1
for explanation of interpretation of statistics and plots.}
\usage{
valCheck(lm, series = FALSE)
}
\arguments{
\item{lm}{A linear model object.}

\item{series}{A boolean variable to indicate whether linear model has a time series component.}
}
\value{
Statistics will be printed to the command line statistics. Plots will be printed.
}
\description{
Check Validity of Linear Model

This function will help determine the validity of a linear model.
The function produces statistics and or plots to assess
equality of variance, normality, multicollinearity, and outliers.
If \code{series=TRUE} then the function will also print a statistic
for independence. See vignette MLReg20_Explanation_And_Theory Section 1
for explanation of interpretation of statistics and plots.
}
\examples{

iris<-as.data.frame(datasets::iris)
iris<-within(iris, {
Setosa<-ifelse(Species=="setosa", 1, 0)
Versicolor<-ifelse(Species=="versicolor", 1, 0)
rm(Species)
})

irislm<-lm(Sepal.Length~., data=iris)
valCheck(irislm)

eu<-as.data.frame(datasets::EuStockMarkets)
form.eu<-"DAX~SMI+CAC+FTSE"
eulm<-lm(form.eu, data=eu)
valCheck(eulm, series=TRUE)

}
