% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valCheck.R
\name{valCheck}
\alias{valCheck}
\title{Check the validity of a linear model}
\usage{
valCheck(lm)
}
\arguments{
\item{lm}{a linear model}
}
\value{
command line statistics
}
\description{
Check the validity of a linear model
}
